StudentModel::login() uses password_hash()
login() prefers to check the provided password against the stored hash
using password_verify(). If the check fails but the stored hash matches
the md5 hash of the provided password, login() creates a new hash using
password_hash() and stores it.

Also changed calls to md5() in StudentController.php to use
password_verify().